+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - ⚠️ Switching from [`rustls-native-certs`](https://crates.io/crates/rustls-native-certs) to [`rustls-platform-verifier`](https://crates.io/crates/rustls-platform-verifier) for loading system certificates.  
     This might imply subtle behavior changes when using `ConnectionConfig::with_encryption_trust_default_cas` or equivalent configurations.
   - ⚠️ Bump MSRV to 1.88.
+ - ⚠️ Many fields in `neo4j::summary::Profile` are now `Option`s to reflect that they might not have been recorded during the Query's execution.
 
 **🧹️ Clean-up**
  - Remove all exposed 3rd party types.